Updating the package.json
Updating the package.json and sate files to allow the Sate system to be
installed globally within node.  For example, running the following in
the satecms directory:

npm install -g .

Will install the sate as a command line argument and put this in the
/usr/local/bin as satecms.  You can then run:

satecms develop

Within a directory with website.json to run sate with all defaults.
There are some issues with this right now, but I'm working through them.